requirements.txt
autopep8==1.7.0
beautifulsoup4==4.11.1
blis==0.7.8
Bottleneck
# catalogue==2.0.8
certifi
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
fast-autocomplete
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
# murmurhash==1.0.8
myst-parser
# networkx==2.8.7
nltk==3.7
numexpr
numpy
packaging
pandas
pathy
Pillow
# preshed
pycodestyle
pydantic
pylev
pyparsing
python-dateutil
python-Levenshtein
pytz
PyYAML==6.0
rake-nltk==1.0.6
regex==2022.9.13
requests==2.28.1
scikit-learn
scipy
segtok==1.5.11
sentence-transformers==2.2.2
sentencepiece==0.1.97
six
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
tabulate==0.9.0
thinc==8.1.3
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.22.2
typer==0.4.2
# typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
yake==0.4.8
